+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/test/utils/TestConstants.java
@@ -1,0 +1,7 @@
+package org.hl7.fhir.r5.test.utils;
+
+import java.nio.file.Paths;
+
+public class TestConstants {
+    public static final java.lang.String TX_CACHE = Paths.get("src","test","resources", "txCache").toAbsolutePath().toString();
+}