+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/cli/utils/Params.java
@@ -37,6 +37,7 @@ public class Params {
   public static final String SCAN = "-scan";
   public static final String TERMINOLOGY = "-tx";
   public static final String TERMINOLOGY_LOG = "-txLog";
+  public static final String TERMINOLOGY_CACHE = "-txCache";
   public static final String LOG = "-log";
   public static final String LANGUAGE = "-language";
   public static final String IMPLEMENTATION_GUIDE = "-ig";
@@ -228,7 +229,13 @@ public class Params {
           throw new Error("Specified -txLog without indicating file");
         else
           cliContext.setTxLog(args[++i]);
-      } else if (args[i].equals(LOG)) {
+      } else if (args[i].equals(TERMINOLOGY_CACHE)) {
+        if (i + 1 == args.length)
+          throw new Error("Specified -txCache without indicating file");
+        else
+          cliContext.setTxCache(args[++i]);
+      }
+      else if (args[i].equals(LOG)) {
         if (i + 1 == args.length)
           throw new Error("Specified -log without indicating file");
         else